+++ NEWS.md
@@ -7,6 +7,7 @@
 * Added `print()`-methods for some more functions, for a clearer output.
 * Revised `r2()` for mixed models (packages **lme4**, **glmmTMB**). The r-squared value should be much more precise now, and reports the marginal and conditional r-squared values.
 * Reduced package dependencies and removed _apaTables_ and _MBESS_ from suggested packages
+* `stanmvreg`-models are now supported by many functions.
 
 ## New functions